feat : 현재 에러있음 메모 수정 세션 수정

+++ backend/src/main/java/com/pj/planbee/service/CalendarService.java
@@ -7,7 +7,7 @@ import com.pj.planbee.dto.CalendarDTO;
 
 public interface CalendarService {
 
-   public double getProgress(String calDate, String sessionId);
+   public int getProgress(String calDate, String sessionId);
    
    public Map<String, Integer> curProgress(String userId);
